moved helper to advice classes. Migrate es-restclient to indy plugin
