+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-5_6/src/main/java/co/elastic/apm/agent/es/restclient/v5_6/ElasticsearchClientAsyncInstrumentation.java
@@ -24,10 +24,9 @@
  */
 package co.elastic.apm.agent.es.restclient.v5_6;
 
-import co.elastic.apm.agent.bci.VisibleForAdvice;
 import co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentation;
 import co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelper;
-import co.elastic.apm.agent.impl.ElasticApmTracer;
+import co.elastic.apm.agent.impl.transaction.AbstractSpan;
 import co.elastic.apm.agent.impl.transaction.Span;
 import co.elastic.apm.agent.sdk.advice.AssignTo;
 import co.elastic.apm.agent.sdk.state.GlobalThreadLocal;
@@ -36,7 +35,6 @@ import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 import org.apache.http.HttpEntity;
-import org.elasticsearch.client.Response;
 import org.elasticsearch.client.ResponseListener;
 
 import javax.annotation.Nullable;
@@ -48,10 +46,6 @@ import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 public class ElasticsearchClientAsyncInstrumentation extends ElasticsearchRestClientInstrumentation {
-
-    public ElasticsearchClientAsyncInstrumentation(ElasticApmTracer tracer) {
-        super(tracer);
-    }
 
     @Override
     public Class<?> getAdviceClass() {
@@ -77,27 +71,25 @@ public class ElasticsearchClientAsyncInstrumentation extends ElasticsearchRestCl
 
 
     public static class ElasticsearchRestClientAsyncAdvice {
-        @VisibleForAdvice
+        public static ElasticsearchRestClientInstrumentationHelper helper = new ElasticsearchRestClientInstrumentationHelper();
         public static final GlobalThreadLocal<Span> spanTls = GlobalThreadLocal.get(ElasticsearchRestClientAsyncAdvice.class, "spanTls");
+
         @AssignTo.Argument(5)
-        @Advice.OnMethodEnter(suppress = Throwable.class)
+        @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
         public static ResponseListener onBeforeExecute(@Advice.Argument(0) String method,
                                                        @Advice.Argument(1) String endpoint,
                                                        @Advice.Argument(3) @Nullable HttpEntity entity,
                                                        @Advice.Argument(5) ResponseListener responseListener) {
-
-            ElasticsearchRestClientInstrumentationHelper<HttpEntity, Response, ResponseListener> helper = esClientInstrHelperManager.getForClassLoaderOfClass(Response.class);
-            if (helper != null) {
-                Span span = helper.createClientSpan(method, endpoint, entity);
-                spanTls.set(span);
-                if (span != null) {
-                    return helper.<ResponseListener>wrapResponseListener(responseListener, span);
-                }
+            AbstractSpan<?> activeSpan = tracer.getActive();
+            Span span = helper.createClientSpan(activeSpan, method, endpoint, entity);
+            spanTls.set(span);
+            if (span != null) {
+                return helper.<ResponseListener>wrapResponseListener(responseListener, span);
             }
             return responseListener;
         }
 
-        @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class)
+        @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class, inline = false)
         public static void onAfterExecute(@Advice.Thrown @Nullable Throwable t) {
             final Span span = spanTls.getAndRemove();
             if (span != null) {
