+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-6_4/src/main/java/co/elastic/apm/agent/es/restclient/v6_4/ElasticsearchClientAsyncInstrumentation.java
@@ -24,10 +24,8 @@
  */
 package co.elastic.apm.agent.es.restclient.v6_4;
 
-import co.elastic.apm.agent.bci.VisibleForAdvice;
 import co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentation;
 import co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelper;
-import co.elastic.apm.agent.impl.ElasticApmTracer;
 import co.elastic.apm.agent.impl.transaction.Span;
 import co.elastic.apm.agent.sdk.advice.AssignTo;
 import co.elastic.apm.agent.sdk.state.GlobalThreadLocal;
@@ -35,9 +33,7 @@ import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
-import org.apache.http.HttpEntity;
 import org.elasticsearch.client.Request;
-import org.elasticsearch.client.Response;
 import org.elasticsearch.client.ResponseListener;
 
 import javax.annotation.Nullable;
@@ -47,10 +43,6 @@ import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 public class ElasticsearchClientAsyncInstrumentation extends ElasticsearchRestClientInstrumentation {
-
-    public ElasticsearchClientAsyncInstrumentation(ElasticApmTracer tracer) {
-        super(tracer);
-    }
 
     @Override
     public Class<?> getAdviceClass() {
@@ -71,25 +63,22 @@ public class ElasticsearchClientAsyncInstrumentation extends ElasticsearchRestCl
     }
 
     public static class ElasticsearchRestClientAsyncAdvice {
-        @VisibleForAdvice
+        public static ElasticsearchRestClientInstrumentationHelper helper = new ElasticsearchRestClientInstrumentationHelper();
         public static final GlobalThreadLocal<Span> spanTls = GlobalThreadLocal.get(ElasticsearchRestClientAsyncAdvice.class, "spanTls");
 
         @AssignTo.Argument(1)
-        @Advice.OnMethodEnter(suppress = Throwable.class)
+        @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
         public static ResponseListener onBeforeExecute(@Advice.Argument(0) Request request,
                                                        @Advice.Argument(1) ResponseListener responseListener) {
-            ElasticsearchRestClientInstrumentationHelper<HttpEntity, Response, ResponseListener> helper = esClientInstrHelperManager.getForClassLoaderOfClass(Request.class);
-            if (helper != null) {
-                Span span = helper.createClientSpan(request.getMethod(), request.getEndpoint(), request.getEntity());
-                if (span != null) {
-                    spanTls.set(span);
-                    return helper.<ResponseListener>wrapResponseListener(responseListener, span);
-                }
+            Span span = helper.createClientSpan(tracer.getActive(), request.getMethod(), request.getEndpoint(), request.getEntity());
+            if (span != null) {
+                spanTls.set(span);
+                return helper.<ResponseListener>wrapResponseListener(responseListener, span);
             }
             return responseListener;
         }
 
-        @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class)
+        @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class, inline = false)
         public static void onAfterExecute(@Advice.Thrown @Nullable Throwable t) {
             final Span span = spanTls.getAndRemove();
             if (span != null) {
