+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-6_4/src/main/java/co/elastic/apm/agent/es/restclient/v6_4/ElasticsearchClientSyncInstrumentation.java
@@ -26,16 +26,13 @@ package co.elastic.apm.agent.es.restclient.v6_4;
 
 import co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentation;
 import co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelper;
-import co.elastic.apm.agent.impl.ElasticApmTracer;
 import co.elastic.apm.agent.impl.transaction.Span;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
-import org.apache.http.HttpEntity;
 import org.elasticsearch.client.Request;
 import org.elasticsearch.client.Response;
-import org.elasticsearch.client.ResponseListener;
 
 import javax.annotation.Nullable;
 
@@ -44,10 +41,6 @@ import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 public class ElasticsearchClientSyncInstrumentation extends ElasticsearchRestClientInstrumentation {
-
-    public ElasticsearchClientSyncInstrumentation(ElasticApmTracer tracer) {
-        super(tracer);
-    }
 
     @Override
     public Class<?> getAdviceClass() {
@@ -67,25 +60,20 @@ public class ElasticsearchClientSyncInstrumentation extends ElasticsearchRestCli
     }
 
     public static class ElasticsearchRestClientSyncAdvice {
-        @Nullable
-        @Advice.OnMethodEnter(suppress = Throwable.class)
-        public static Span onBeforeExecute(@Advice.Argument(0) Request request) {
+        public static ElasticsearchRestClientInstrumentationHelper helper = new ElasticsearchRestClientInstrumentationHelper();
 
-            ElasticsearchRestClientInstrumentationHelper<HttpEntity, Response, ResponseListener> helper =
-                esClientInstrHelperManager.getForClassLoaderOfClass(Request.class);
-            if (helper == null) {
-                return null;
-            }
-            return helper.createClientSpan(request.getMethod(), request.getEndpoint(), request.getEntity());
+        @Nullable
+        @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
+        public static Object onBeforeExecute(@Advice.Argument(0) Request request) {
+            return helper.createClientSpan(tracer.getActive(), request.getMethod(), request.getEndpoint(), request.getEntity());
         }
 
-        @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class)
+        @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class, inline = false)
         public static void onAfterExecute(@Advice.Return @Nullable Response response,
-                                          @Advice.Enter @Nullable Span span,
+                                          @Advice.Enter @Nullable Object spanObj,
                                           @Advice.Thrown @Nullable Throwable t) {
-            ElasticsearchRestClientInstrumentationHelper<HttpEntity, Response, ResponseListener> helper =
-                esClientInstrHelperManager.getForClassLoaderOfClass(Request.class);
-            if (helper != null && span != null) {
+            if (spanObj instanceof Span) {
+                Span span = (Span) spanObj;
                 try {
                     helper.finishClientSpan(response, span, t);
                 } finally {
