+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-common/src/main/java/co/elastic/apm/agent/es/restclient/ResponseListenerWrapper.java
@@ -11,9 +11,9 @@
  * the Apache License, Version 2.0 (the "License"); you may
  * not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -33,13 +33,13 @@ import javax.annotation.Nullable;
 
 public class ResponseListenerWrapper implements ResponseListener, Recyclable {
 
-    private ElasticsearchRestClientInstrumentationHelperImpl helper;
+    private ElasticsearchRestClientInstrumentationHelper helper;
     @Nullable
     private ResponseListener delegate;
     @Nullable
     private volatile Span span;
 
-    ResponseListenerWrapper(ElasticsearchRestClientInstrumentationHelperImpl helper) {
+    ResponseListenerWrapper(ElasticsearchRestClientInstrumentationHelper helper) {
         this.helper = helper;
     }
