+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-common/src/test/java/co/elastic/apm/agent/es/restclient/AbstractEsClientInstrumentationTest.java
@@ -40,9 +40,9 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 
-import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelperImpl.ELASTICSEARCH;
-import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelperImpl.SPAN_ACTION;
-import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelperImpl.SPAN_TYPE;
+import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelper.ELASTICSEARCH;
+import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelper.SPAN_ACTION;
+import static co.elastic.apm.agent.es.restclient.ElasticsearchRestClientInstrumentationHelper.SPAN_TYPE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public abstract class AbstractEsClientInstrumentationTest extends AbstractInstrumentationTest {
